[DEV-10286] allow name change on toggles (#1948)
* [DEV-10286] allow name change on toggles

* [DEV-10286] updates for widget styling and toggle name status

* [DEV-10286] update dropdown

* [DEV-10286] change type on ColumnProps
